Move scraper tests insidde wsf_scraping
