--- pipeline/reach-scraper/tests/mock_sites/gov/1.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<!--[if lt IE 9]><html class="lte-ie8" lang="en"><![endif]--><!--[if gt IE 8]><!--><html lang="en">
-<!--<![endif]-->
-  <head>
-    <meta charset="utf-8">
-    <title>Policy papers and consultations - GOV.UK</title>
-</head>
-
-  <body class="">
-
-  <div class="facet-tags" data-module="track-click">
-      <div class="facet-tags__group">
-          <div class="facet-tags__wrapper">
-            <p class="facet-tags__preposition">Of Type</p>
-            <div class="facet-tag">
-              <p class="facet-tag__text">Policy papers</p>
-              <button type="button" class="facet-tag__remove" aria-label="Remove filter Policy papers" data-module="remove-filter-link" data-track-label="Policy papers" data-facet="content_store_document_type" data-value="policy_papers" data-name="">✕</button>
-            </div>
-          </div>
-      </div>
-  </div>
-
-          </div>
-        </div>
-
-        <div class="govuk-caption-l live-search-loading-message" id="js-loading-message"></div>
-
-        <div id="js-sort-options">
-            <div class="govuk-form-group">
-    <label for="order" class="result-region-header__sort-label govuk-label">Sort by</label>
-    <select class="js-order-results govuk-select" name="order" id="order" aria-controls="js-search-results-info" data-default-sort-option="updated-newest" data-relevance-sort-option="relevance" data-module="track-select-change">
-        <option value="most-viewed" data-track-category="dropDownClicked" data-track-action="clicked" data-track-label="Most viewed">Most viewed</option>
-        <option value="relevance" disabled data-track-category="dropDownClicked" data-track-action="clicked" data-track-label="Relevance">Relevance</option>
-        <option value="updated-newest" selected data-track-category="dropDownClicked" data-track-action="clicked" data-track-label="Updated (newest)">Updated (newest)</option>
-        <option value="updated-oldest" data-track-category="dropDownClicked" data-track-action="clicked" data-track-label="Updated (oldest)">Updated (oldest)</option>
-    </select>
-  </div>
-
-        </div>
-
-        <div id="js-results">
-            <div class="finder-results js-finder-results" data-module="track-click">
-        <ol class="gem-c-document-list gem-c-document-list--no-underline">
-
-      <li class="gem-c-document-list__item  ">
-        <a data-ecommerce-path="/government/publications/continuing-the-uks-trade-relationship-with-south-korea-parliamentary-report" data-ecommerce-content-id="aa69bee3-7cb1-409e-9dad-bb2c1d083208" data-ecommerce-row="1" data-track-category="navFinderLinkClicked" data-track-action="Policy papers and consultations.1" data-track-label="/government/publications/continuing-the-uks-trade-relationship-with-south-korea-parliamentary-report" data-track-options="{&quot;dimension28&quot;:20,&quot;dimension29&quot;:&quot;Continuing the UK's trade relationship with South Korea: parliamentary report&quot;}" class="gem-c-document-list__item-title  " href="/government/publications/continuing-the-uks-trade-relationship-with-south-korea-parliamentary-report">Continuing the UK's trade relationship with South Korea: parliamentary report</a>
-
-
-          <p class="gem-c-document-list__item-description">This report explains the government’s approach to maintaining continuity of the trade relationship between the UK and South Korea after Brexit.</p>
-
-          <ul class="gem-c-document-list__item-metadata">
-              <li class="gem-c-document-list__attribute">
-                  From: Department for International Trade
-              </li>
-              <li class="gem-c-document-list__attribute">
-                  Updated: <time datetime="2019-09-09">9 September 2019</time>
-              </li>
-          </ul>
-
-      </li>
-
-      <li class="gem-c-document-list__item  ">
-        <a data-ecommerce-path="/government/publications/investing-in-women-code" data-ecommerce-content-id="8e39770b-686e-4bb0-bb65-25bed937b5cb" data-ecommerce-row="1" data-track-category="navFinderLinkClicked" data-track-action="Policy papers and consultations.2" data-track-label="/government/publications/investing-in-women-code" data-track-options='{"dimension28":20,"dimension29":"Investing in Women Code"}' class="gem-c-document-list__item-title  " href="/government/publications/investing-in-women-code">Investing in Women Code</a>
-
-
-          <p class="gem-c-document-list__item-description">The Investing in Women Code is a commitment by financial services firms to improving female entrepreneurs’ access to tools, resources and finance.</p>
-
-          <ul class="gem-c-document-list__item-metadata">
-              <li class="gem-c-document-list__attribute">
-                  From: HM Treasury
-              </li>
-              <li class="gem-c-document-list__attribute">
-                  Updated: <time datetime="2019-09-09">9 September 2019</time>
-              </li>
-          </ul>
-
-      </li>
-    </ol>
-
-  </div>
-</div>
-</body>
-</html>
