--- pipeline/reach-scraper/tests/mock_sites/nice/1.html
@@ -1,64 +1,0 @@
-
-<!DOCTYPE html>
-<html lang="en-GB">
-    <head>
-        <title>Search Results</title>
-<body data-spy="scroll" data-target=".aside-primary" class=" " >
-<!-- Google Tag Manager -->
-
-            <div class="media">
-                <div class="media-body">
-                    <h4 class="media-heading">
-
-                            <a href="/guidance/cg110">Pregnancy and complex social factors: a model for service provision for pregnant women with complex social factors <span class="muted">(CG110)</span></a>
-
-                    </h4>
-
-                        <p> This guideline covers antenatal care for all pregnant women with complex social factors (particularly alcohol or drug misuse, recent migrant or asylum seeker status, difficulty reading or speaking English, aged under 20, domestic abuse). It offers advice on improving access to care, maintaining contact with antenatal carers, and additional information and support for these women. </p>
-
-                    <div class="row">
-		                    <span class="documenttype">Clinical guideline</span>
-
-                                <span class="publisheddate">
-                                    <text>Published&nbsp;</text>September 2010
-                                </span>
-
-                    </div>
-
-                    <div class="result-footer grid5">
-                            <a data-toggle="collapse" data-target="#result-0" class="pull-right btn btn-info">More <i class="icon-caret-down"></i></a>
-                                <div id="result-0" class="collapse over4">
-
-
-
-    <ul class="unstyled grid2">
-                <li><a href="/guidance/cg110">Overview</a></li>
-                <li><a href="/guidance/cg110/chapter/Introduction">Introduction</a></li>
-                <li><a href="/guidance/cg110/chapter/Woman-centred-care">Woman-centred care</a></li>
-                <li><a href="/guidance/cg110/chapter/Key-priorities-for-implementation">Key priorities for implementation</a></li>
-                <li><a href="/guidance/cg110/chapter/1-Guidance">1 Guidance</a></li>
-                <li><a href="/guidance/cg110/chapter/2-Research-recommendations">2 Research recommendations</a></li>
-                <li><a href="/guidance/cg110/chapter/Update-information">Update information</a></li>
-    </ul>
-
-                                </div>
-                    </div>
-                </div>
-            </div>
-
-    <div class="pager pager-centered">
-        <ul>
-                <li class="disabled"><span>Previous</span></li>
-                                        <li class="active"><span>1</span></li>
-                <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=2&amp;ps=50&amp;sp=on">2</a></li>
-                <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=3&amp;ps=50&amp;sp=on">3</a></li>
-                <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=4&amp;ps=50&amp;sp=on">4</a></li>
-                <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=5&amp;ps=50&amp;sp=on">5</a></li>
-                <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=6&amp;ps=50&amp;sp=on">6</a></li>
-                <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=7&amp;ps=50&amp;sp=on">7</a></li>
-                            <li><a href="Search?om=[{&quot;gst&quot;:[&quot;Published&quot;]},{&quot;ndt&quot;:[&quot;Guidance&quot;]}]&amp;pa=2&amp;ps=50&amp;sp=on">Next</a></li>
-        </ul>
-    </div>
-</body>
-
-</html>
