--- pipeline/reach-scraper/tests/mock_sites/nice/2.html
@@ -1,84 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-GB" class="no-js">
-<head>
-    <title>StoneChecker for kidney stone evaluation | Guidance and guidelines | NICE</title>
-</head>
-<body data-spy="scroll" data-target=".aside-primary" class="">
-
-    <div id="printpane"></div>
-
-            <header class="_ page-header">
-                <div class="zone zone-breadcrumb">
-
-                    <div class="container" id="global-breadcrumb">
-
-
-
-                        <h1>StoneChecker for kidney stone evaluation</h1>
-
-    <div class="product-heading  ">
-    <div class="product-heading">
-        <h4 class="product-title">
-            <span class="prod-title">Medtech innovation briefing [MIB171]</span><wbr>
-                <span class="published-date">
-                    Published date:
-                        <time datetime="2019-02-01">February 2019</time>
-                </span>
-                    </h4>
-    </div>
-
-</div>
-                    </div>
-                </div>
-            </header>
-        <div class="_ layout  ">
-            <div id="toTop"><i aria-hidden="true" class="icon-nice-chevron-up"></i></div>
-            <div class="_">
-                <div class="content">
-
-
-
-                    <div class="zone zone-content">
-
-
-
-
-<div id="guidance-tabs-container" class="tabbable">
-
-                </div>
-            <div class="span9 tab-header">
-
-<div class="share-links navbar pull-right">
-    <div class="navbar-inner">
-        <ul class="unstyled">
-            <li class="dropdown">
-
-                    <a href="#" id="nice-share" class="btn-sharelink dropdown-toggle" data-toggle="dropdown" title="Share">
-                        <span>Share</span>
-                    </a>
-            </li>
-                <li class="dropdown">
-                        <a href="#" id="nice-download" class="btn-sharelink dropdown-toggle" data-toggle="dropdown" title="Download">
-                            <span>Download</span>
-                        </a>
-                        <ul class="dropdown-menu pull-right" role="menu" aria-labelledby="nice-download">
-                            <li class="nav-header"><b>Save</b></li>
-                                <li>
-                                    <a class="track-event" target="_blank" href="/guidance/mib171/resources/stonechecker-for-kidney-stone-evaluation-pdf-2285963632524229"
-                                       data-track="guidancedownload" data-track-category="guidance download" data-track-action="clicked" data-track-label="/guidance/mib171/resources/stonechecker-for-kidney-stone-evaluation-pdf-2285963632524229">Save as PDF</a>
-                                </li>
-                                <li>
-                                    <a class="track-event" target="_blank" href="/guidance/mib171/resources/stonechecker-for-kidney-stone-evaluation-epub-2285963632524218"
-                                       data-track="guidancedownload" data-track-category="guidance download" data-track-action="clicked" data-track-label="/guidance/mib171/resources/stonechecker-for-kidney-stone-evaluation-epub-2285963632524218">Save as ePub</a>
-                                </li>
-                                <li>
-                                    <a class="track-event" target="_blank" href="/guidance/mib171/resources/stonechecker-for-kidney-stone-evaluation-ebook-2285963632524226"
-                                       data-track="guidancedownload" data-track-category="guidance download" data-track-action="clicked" data-track-label="/guidance/mib171/resources/stonechecker-for-kidney-stone-evaluation-ebook-2285963632524226">Save as eBook</a>
-                                </li>
-                        </ul>
-                </li>
-        </ul>
-    </div>
-</div>
-</body>
-</html>
