--- pipeline/reach-scraper/tests/mock_sites/parliament/1.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html class="no-js" lang="en"> <!--<![endif]-->
-    <head>
-        <title>Search results</title></head>
-    <body class="main">
-        <ul id="results">
-                    <li>
-                        <h4>
-                            <a href="http://www.parliament.uk/written-questions-answers-statements/written-question/commons/2019-01-29/214040">
-                                Motor Vehicles: Insurance
-                            </a>
-                        </h4>
-                        <p>
-                            &lt;p&gt;To ask the Secretary of State for Transport, what steps his Department is taking to support British citizens driving in EU countries after 29 March 2019 who do not hold a green card.&lt;/p&gt;
-                        </p>
-                        <p class="resultdetails">
-06 Feb 2019                                        <span> | </span>
-Written questions                                        <span> | </span>
-House of Commons                                        <span> | </span>
-214040
-                        </p>
-                    </li>
-            </ul>
-    </body>
-</html>
