--- pipeline/reach-scraper/tests/mock_sites/parliament/2.html
@@ -1,10 +1,0 @@
-
-<!DOCTYPE html
-  PUBLIC "-//W3C//DTD XHTML+RDFa 1.0//EN" "http://www.w3.org/MarkUp/DTD/xhtml-rdfa-1.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" xmlns:dct="http://purl.org/dc/terms/" lang="en" xml:lang="en">
-<head>
-    <title>The Animal Feed (Amendment) (EU Exit) Regulations 2019</title>
-</head>
-<body>
-    <a href="/ukdsi/2019/9780111180389/contents/data.pdf" target="_blank" class="pdfLink">PDF<span class="accessibleText"> table of contents</span></a>
-</body></html>
