--- pipeline/reach-scraper/tests/mock_sites/unicef/1.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US" class="no-js" dir="ltr">
-<head>
-<title>Publications Archives - UNICEF DATA</title>
-</head>
-<body class="archive tax-resource-type term-publications term-18 template--block ">
-<div class="block--card__body__text">
-<h3>
-<a href="https://data.unicef.org/resources/including-everyone-strengthening-the-collection-and-use-of-data-about-persons-with-disabilities-in-humanitarian-situations/" title="Including Everyone: Strengthening the collection and use of data about persons with disabilities in humanitarian situations">Including Everyone: Strengthening the collection and use of data about persons with disabilities in humanitarian situations</a>
-</h3>
-</div>
-<nav class="navigation pagination" role="navigation" aria-label="Posts">
-<h2 class="screen-reader-text">Posts navigation</h2>
-<div class="nav-links"><span aria-current="page" class="page-numbers current"><span class="screen-reader-text">Page </span>1</span>
-<a class="page-numbers" href="https://data.unicef.org/resources/resource-type/publications/page/2/"><span class="screen-reader-text">Page </span>2</a>
-<span class="page-numbers dots">&hellip;</span>
-<a class="page-numbers" href="https://data.unicef.org/resources/resource-type/publications/page/10/"><span class="screen-reader-text">Page </span>10</a>
-<a class="next page-numbers" href="https://data.unicef.org/resources/resource-type/publications/page/2/">Next page</a></div>
-</nav>
-</body>
-</html>
