--- pipeline/reach-scraper/tests/mock_sites/unicef/2.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en-US" prefix="og: http://ogp.me/ns#" class="no-js" dir="ltr">
-<head>
-    <title>A Right to be Heard - Listening to children and young people on the move - UNICEF DATA</title>
-</head>
-<body class="post-template-default single single-post postid-9761 single-format-standard template--block ">
-<p style="text-align: center;"><a class="button button--size-large button--color-blue " title="Migration, Refugees and migrants | A Right to be Heard – Listening to children and young people on the move | English | PDF" href="https://data.unicef.org/wp-content/uploads/2018/12/A_right_to_be_heard_youthpoll.pdf" target="_blank" rel="noopener" data-category="Resources" data-action="Download">Download</a></p>
-</body>
-</html>
