--- pipeline/reach-scraper/tests/mock_sites/who/1.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<title>IRIS Home</title>
-</head><body>
-<ul>
-<li class="ds-artifact-item even">
-<div class="item-wrapper row">
-<div class="col-sm-9">
-<div class="artifact-description">
-<h4 class="artifact-title">
-<a href="./tests/mock_sites/2.html">Air pollution control: report on an inter-regional seminar convened by the World Health Organization in collaboration with the Government of the USSR</a><span class="Z3988" title="ctx_ver=Z39.88-2004&amp;rft_val_fmt=info%3Aofi%2Ffmt%3Akev%3Amtx%3Adc&amp;rft_id=EP%2F68.2&amp;rft_id=http%3A%2F%2Fwww.who.int%2Firis%2Fhandle%2F10665%2F279983&amp;rfr_id=info%3Asid%2Fdspace.org%3Arepository&amp;">
-                    ﻿
-                </span>
-</h4>
-<div class="artifact-info">
-<span class="author h4"><small><span>World Health Organization</span></small></span> <span class="govdoc-date h4"><small>(&lrm;<span class="govdoc">EP/68.2</span>, <span class="date">1967</span>)&lrm;</small></span>
-</div>
-</div>
-</div>
-</div>
-</li>
-</ul>
-</body></html>
