--- pipeline/reach-scraper/tests/mock_sites/who/2.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html><html class="no-js" lang="en"> <!--<![‎endif]‎-->
-            <head>
-<title>Air pollution control: report on an inter-regional seminar convened by the World Health Organization in collaboration with the Government of the USSR</title>
-</head><body>
-<h2 class="page-header first-page-header">Air pollution control: report on an inter-regional seminar convened by the World Health Organization in collaboration with the Government of the USSR</h2>
-<div class="ds-table-responsive">
-<table class="ds-includeSet-table detailtable table table-striped table-hover">
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.contributor.author</td><td class="word-break">World Health Organization</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.coverage.spatial</td><td class="word-break">Geneva</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.date.accessioned</td><td class="word-break">2019-02-04T15:02:50Z</td><td></td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.date.available</td><td class="word-break">2019-02-04T15:02:50Z</td><td></td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.date.issued</td><td class="word-break">1967</td><td></td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.identifier.govdoc</td><td class="word-break">EP/68.2</td><td></td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.identifier.uri</td><td class="word-break">http://www.who.int/iris/handle/10665/279983</td><td></td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.description</td><td class="word-break">X-173 p.</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.language.iso</td><td class="word-break">en</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.publisher</td><td class="word-break">World Health Organization</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.subject.mesh</td><td class="word-break">Air Pollution</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.subject.mesh</td><td class="word-break">Meeting Abstracts</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.subject.mesh</td><td class="word-break">USSR</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row even ">
-<td class="label-cell">dc.title</td><td class="word-break">Air pollution control: report on an inter-regional seminar convened by the World Health Organization in collaboration with the Government of the USSR</td><td>en_US</td>
-</tr>
-
-<tr class="ds-table-row odd ">
-<td class="label-cell">dc.type</td><td class="word-break">Technical documents</td><td>en_US</td>
-</tr>
-
-</table>
-</div>
-<span class="Z3988" title="ctx_ver=Z39.88-2004&amp;rft_val_fmt=info%3Aofi%2Ffmt%3Akev%3Amtx%3Adc&amp;rft_id=EP%2F68.2&amp;rft_id=http%3A%2F%2Fwww.who.int%2Firis%2Fhandle%2F10665%2F279983&amp;rfr_id=info%3Asid%2Fdspace.org%3Arepository&amp;">
-            ﻿
-        </span>
-<h3>Files in this item</h3>
-<div class="file-list">
-<div class="file-wrapper row">
-<div class="col-xs-6 col-sm-3">
-<div class="thumbnail">
-<a class="image-link" href="/iris/bitstream/handle/10665/279983/EP-68.2-eng.pdf?sequence=1&isAllowed=y"><img alt="Thumbnail" src="/iris/bitstream/handle/10665/279983/EP-68.2-eng.pdf.jpg?sequence=8&isAllowed=y" style="max-width: 130px; max-height: 190px;"></a>
-</div>
-</div>
-<div class="col-xs-6 col-sm-7">
-<dl class="file-metadata dl-horizontal">
-<dt>Name:</dt>
-<dd class="word-break" title="EP-68.2-eng.pdf">EP-68.2-eng.pdf</dd>
-<dt>Size:</dt>
-<dd class="word-break">7.758Mb</dd>
-<dt>Format:</dt>
-<dd class="word-break">PDF</dd>
-</dl>
-</div>
-<div class="file-link col-xs-6 col-xs-offset-6 col-sm-2 col-sm-offset-0">
-<a href="foo.pdf">View/<wbr></wbr>Open</a>
-</body></html>
