--- pipeline/reach-scraper/tests/xml/test_xml.xml
@@ -1,20 +1,0 @@
-<pdf2xml producer="poppler" version="0.62.0">
-<page number="1" position="absolute" top="0" left="0" height="1262" width="892">
-<fontspec id="0" size="12" family="Times" color="#000000"/>
-<fontspec id="1" size="25" family="Times" color="#000000"/>
-<text top="141" left="446" width="475" height="25" font="1">Test Page 1</text>
-<text top="173" left="446" width="475" height="12" font="0"><b>All bold line.</b></text>
-<text top="205" left="446" width="475" height="12" font="0">Partly <b>bold</b> line.</text>
-<text top="173" left="446" width="475" height="12" font="0"><i>All italic line.</i></text>
-<text top="205" left="446" width="475" height="12" font="0">Partly <i>italic</i> line.</text>
-</page>
-<page number="2" position="absolute" top="0" left="0" height="1262" width="892">
-<fontspec id="0" size="12" family="Times" color="#000000"/>
-<fontspec id="1" size="25" family="Times" color="#000000"/>
-<text top="141" left="446" width="475" height="25" font="1">TestPage 2</text>
-<text top="173" left="446" width="475" height="12" font="0"><b>All bold line</b></text>
-<text top="205" left="446" width="475" height="12" font="0">Partly <b>bold</b> line.</text>
-<text top="173" left="446" width="475" height="12" font="0"><i>All italic line</i></text>
-<text top="205" left="446" width="475" height="12" font="0">Partly <i>italic</i> line.</text>
-</page>
-</pdf2xml>
