+++ pipeline/reach-scraper/wsf_scraping/tests/test_nice_spider.py
@@ -5,7 +5,7 @@ from scrapy.http import Response, Request, HtmlResponse, TextResponse
 from scrapy.utils.project import get_project_settings
 from wsf_scraping.spiders.nice_spider import NiceSpider
 
-from .common import get_path, TEST_PDF
+from tests.common import get_path, TEST_PDF
 
 NICE_2_HTML = get_path('mock_sites/nice/2.html')
