+++ pipeline/reach-scraper/wsf_scraping/tests/test_unicef_spider.py
@@ -3,7 +3,7 @@ from scrapy.http import Response, Request, HtmlResponse
 from scrapy.utils.project import get_project_settings
 from wsf_scraping.spiders.unicef_spider import UnicefSpider
 
-from .common import get_path, TEST_PDF
+from tests.common import get_path, TEST_PDF
 
 
 class Crawler:
