+++ pipeline/reach-scraper/wsf_scraping/tests/test_who_spider.py
@@ -5,7 +5,7 @@ from scrapy.http import Response, Request, HtmlResponse
 from scrapy.utils.project import get_project_settings
 from wsf_scraping.spiders.who_iris_spider import WhoIrisSpider
 
-from .common import get_path, TEST_PDF
+from tests.common import get_path, TEST_PDF
 
 
 class Crawler:
